Fix bug not getting non-default user mfa token
This happens when you have a 'default' user profile but you wish to get
a temporary session with another user profile.  The commit shows clearly
that in the past the 'default' role was used to get an sts client and
then it applied an unrecognised MFA serial to it.

Here the fix is to ensure the sts client session is under the correct
user before getting its temp tokens.